fix for code reviews
