+++ MonMaraicher/Assets.xcassets/farmer1.imageset/Contents.json
@@ -2,16 +2,7 @@
   "images" : [
     {
       "filename" : "uriel-soberanes-71yNGYk9Oyg-unsplash.jpg",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "3x"
+      "idiom" : "universal"
     }
   ],
   "info" : {
