+++ MonMaraicher/Assets.xcassets/farmer2.imageset/Contents.json
@@ -2,16 +2,7 @@
   "images" : [
     {
       "filename" : "gabriel-jimenez-jin4W1HqgL4-unsplash.jpg",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "3x"
+      "idiom" : "universal"
     }
   ],
   "info" : {
