+++ MonMaraicher/Assets.xcassets/farmer3.imageset/Contents.json
@@ -2,16 +2,7 @@
   "images" : [
     {
       "filename" : "markus-spiske-sFydXGrt5OA-unsplash.jpg",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "3x"
+      "idiom" : "universal"
     }
   ],
   "info" : {
